+++ src/tlsrequest.c
@@ -112,7 +112,7 @@ iDefineObjectConstructionArgs(CachedSession,
                               sess, cert)
 
 static void reuse_CachedSession(const iCachedSession *d, SSL *ssl) {
-    BIO *buf = BIO_new_mem_buf(constData_Block(&d->pemSession), size_Block(&d->pemSession));
+    BIO *buf = BIO_new_mem_buf(constData_Block(&d->pemSession), (int) size_Block(&d->pemSession));
     SSL_SESSION *sess = NULL;
     PEM_read_bio_SSL_SESSION(buf, &sess, NULL, NULL);
     if (sess) {
@@ -356,7 +356,7 @@ static iTlsCertificate *newX509Chain_TlsCertificate_(X509 *cert, STACK_OF(X509) 
 
 iTlsCertificate *newPem_TlsCertificate(const iString *pem) {
     iTlsCertificate *d = new_TlsCertificate();
-    BIO *buf = BIO_new_mem_buf(cstr_String(pem), size_String(pem));
+    BIO *buf = BIO_new_mem_buf(cstr_String(pem), (int) size_String(pem));
     PEM_read_bio_X509(buf, &d->cert, NULL /* no passphrase callback */, "" /* empty passphrase */);
     BIO_free(buf);
     return d;
@@ -364,7 +364,7 @@ iTlsCertificate *newPem_TlsCertificate(const iString *pem) {
 
 iTlsCertificate *newPemKey_TlsCertificate(const iString *certPem, const iString *keyPem) {
     iTlsCertificate *d = newPem_TlsCertificate(certPem);
-    BIO *buf = BIO_new_mem_buf(cstr_String(keyPem), size_String(keyPem));
+    BIO *buf = BIO_new_mem_buf(cstr_String(keyPem), (int) size_String(keyPem));
     PEM_read_bio_PrivateKey(buf, &d->pkey, NULL, "");
     BIO_free(buf);
     return d;
@@ -384,7 +384,7 @@ static void add_X509Name_(X509_NAME *name, const char *id, enum iTlsCertificateN
     const iString *str = findName_(names, type);
     if (str) {
         X509_NAME_add_entry_by_txt(
-            name, id, MBSTRING_UTF8, constData_Block(&str->chars), size_String(str), -1, 0);
+            name, id, MBSTRING_UTF8, constData_Block(&str->chars), (int) size_String(str), -1, 0);
     }
 }
 
@@ -396,7 +396,7 @@ static void addDomain_X509Name_(X509_NAME *name, enum iTlsCertificateNameType ty
         iRangecc comp = iNullRange;
         while (nextSplit_Rangecc(range, ".", &comp)) {
             X509_NAME_add_entry_by_txt(
-                name, "DC", MBSTRING_UTF8, (const void *) comp.start, size_Range(&comp), -1, 0);
+                name, "DC", MBSTRING_UTF8, (const void *) comp.start, (int) size_Range(&comp), -1, 0);
         }
     }
 }
@@ -803,7 +803,7 @@ static iBool encrypt_TlsRequest_(iTlsRequest *d) {
         return iFalse;
     }
     while (!isEmpty_Block(&d->sending)) {
-        int n = SSL_write(d->ssl, constData_Block(&d->sending), size_Block(&d->sending));
+        int n = SSL_write(d->ssl, constData_Block(&d->sending), (int) size_Block(&d->sending));
         enum iSSLResult status = sslResult_TlsRequest_(d, n);
         if (n > 0) {
             remove_Block(&d->sending, 0, n);
@@ -919,7 +919,7 @@ static int processIncoming_TlsRequest_(iTlsRequest *d, const char *src, size_t l
     int n;
     do {
         if (len > 0) {
-            n = BIO_write(d->rbio, src, len);
+            n = BIO_write(d->rbio, src, (int) len);
             if (n <= 0) {
                 return -1; /* assume bio write failure is unrecoverable */
             }
